fixed macros in .include'd files
If a macro is encountered, pass1() is called recursively. This
overwrites the global incl_sp, which is the stack pointer for
the include stack. So using a macro in anything but the top-level
source file ended assembly after this file.

The fix saves the value of incl_sp across the recursive call.